Move server context implementation to grpc_impl namespace and typedef the ref to it
