+++ include/grpcpp/server_impl.h
@@ -44,7 +44,6 @@ struct grpc_server;
 
 namespace grpc {
 class AsyncGenericService;
-class ServerContext;
 
 namespace internal {
 class ExternalConnectionAcceptorImpl;
@@ -54,6 +53,7 @@ class ExternalConnectionAcceptorImpl;
 
 namespace grpc_impl {
 class HealthCheckServiceInterface;
+class ServerContext;
 class ServerInitializer;
 
 /// Represents a gRPC server.
@@ -82,9 +82,9 @@ class Server : public grpc::ServerInterface, private grpc::GrpcLibraryCodegen {
     /// Called before server is created.
     virtual void UpdateArguments(grpc::ChannelArguments* args) {}
     /// Called before application callback for each synchronous server request
-    virtual void PreSynchronousRequest(grpc::ServerContext* context) = 0;
+    virtual void PreSynchronousRequest(grpc_impl::ServerContext* context) = 0;
     /// Called after application callback for each synchronous server request
-    virtual void PostSynchronousRequest(grpc::ServerContext* context) = 0;
+    virtual void PostSynchronousRequest(grpc_impl::ServerContext* context) = 0;
     /// Called before server is started.
     virtual void PreServerStart(Server* server) {}
     /// Called after a server port is added.
