+++ src/cpp/ext/filters/census/grpc_plugin.h
@@ -25,8 +25,11 @@
 #include "include/grpcpp/opencensus.h"
 #include "opencensus/stats/stats.h"
 
-namespace grpc {
+namespace grpc_impl {
 class ServerContext;
+}
+
+namespace grpc {
 
 // The tag keys set when recording RPC stats.
 ::opencensus::stats::TagKey ClientMethodTagKey();
