+++ src/cpp/server/server_context.cc
@@ -16,7 +16,7 @@
  *
  */
 
-#include <grpcpp/server_context.h>
+#include <grpcpp/impl/codegen/server_context_impl.h>
 #include <grpcpp/support/server_callback.h>
 
 #include <algorithm>
@@ -28,23 +28,25 @@
 #include <grpc/load_reporting.h>
 #include <grpc/support/alloc.h>
 #include <grpc/support/log.h>
-#include <grpcpp/completion_queue.h>
 #include <grpcpp/impl/call.h>
+#include <grpcpp/impl/codegen/completion_queue_impl.h>
 #include <grpcpp/support/time.h>
 
 #include "src/core/lib/gprpp/ref_counted.h"
 #include "src/core/lib/gprpp/sync.h"
 #include "src/core/lib/surface/call.h"
 
-namespace grpc {
+namespace grpc_impl {
 
 // CompletionOp
 
-class ServerContext::CompletionOp final : public internal::CallOpSetInterface {
+class ServerContext::CompletionOp final
+    : public ::grpc::internal::CallOpSetInterface {
  public:
   // initial refs: one in the server context, one in the cq
   // must ref the call before calling constructor and after deleting this
-  CompletionOp(internal::Call* call, internal::ServerReactor* reactor)
+  CompletionOp(::grpc::internal::Call* call,
+               ::grpc::internal::ServerReactor* reactor)
       : call_(*call),
         reactor_(reactor),
         has_tag_(false),
@@ -67,7 +69,7 @@ class ServerContext::CompletionOp final : public internal::CallOpSetInterface {
     }
   }
 
-  void FillOps(internal::Call* call) override;
+  void FillOps(::grpc::internal::Call* call) override;
 
   // This should always be arena allocated in the call, so override delete.
   // But this class is not trivially destructible, so must actually call delete
@@ -149,8 +151,8 @@ class ServerContext::CompletionOp final : public internal::CallOpSetInterface {
     return finalized_ ? (cancelled_ != 0) : false;
   }
 
-  internal::Call call_;
-  internal::ServerReactor* const reactor_;
+  ::grpc::internal::Call call_;
+  ::grpc::internal::ServerReactor* const reactor_;
   bool has_tag_;
   void* tag_;
   void* core_cq_tag_;
@@ -160,7 +162,7 @@ class ServerContext::CompletionOp final : public internal::CallOpSetInterface {
   int cancelled_;  // This is an int (not bool) because it is passed to core
   std::function<void()> cancel_callback_;
   bool done_intercepting_;
-  internal::InterceptorBatchMethodsImpl interceptor_methods_;
+  ::grpc::internal::InterceptorBatchMethodsImpl interceptor_methods_;
 };
 
 void ServerContext::CompletionOp::Unref() {
@@ -171,7 +173,7 @@ void ServerContext::CompletionOp::Unref() {
   }
 }
 
-void ServerContext::CompletionOp::FillOps(internal::Call* call) {
+void ServerContext::CompletionOp::FillOps(::grpc::internal::Call* call) {
   grpc_op ops;
   ops.op = GRPC_OP_RECV_CLOSE_ON_SERVER;
   ops.data.recv_close_on_server.cancelled = &cancelled_;
@@ -227,7 +229,7 @@ bool ServerContext::CompletionOp::FinalizeResult(void** tag, bool* status) {
   }
   /* Add interception point and run through interceptors */
   interceptor_methods_.AddInterceptionHookPoint(
-      experimental::InterceptionHookPoints::POST_RECV_CLOSE);
+      ::grpc::experimental::InterceptionHookPoints::POST_RECV_CLOSE);
   if (interceptor_methods_.RunInterceptors()) {
     /* No interceptors were run */
     if (has_tag_) {
@@ -292,9 +294,9 @@ void ServerContext::Clear() {
   }
 }
 
-void ServerContext::BeginCompletionOp(internal::Call* call,
-                                      std::function<void(bool)> callback,
-                                      internal::ServerReactor* reactor) {
+void ServerContext::BeginCompletionOp(
+    ::grpc::internal::Call* call, std::function<void(bool)> callback,
+    ::grpc::internal::ServerReactor* reactor) {
   GPR_ASSERT(!completion_op_);
   if (rpc_info_) {
     rpc_info_->Ref();
@@ -313,8 +315,8 @@ void ServerContext::BeginCompletionOp(internal::Call* call,
   call->PerformOps(completion_op_);
 }
 
-internal::CompletionQueueTag* ServerContext::GetCompletionOpTag() {
-  return static_cast<internal::CompletionQueueTag*>(completion_op_);
+::grpc::internal::CompletionQueueTag* ServerContext::GetCompletionOpTag() {
+  return static_cast<::grpc::internal::CompletionQueueTag*>(completion_op_);
 }
 
 void ServerContext::AddInitialMetadata(const grpc::string& key,
@@ -328,7 +330,7 @@ void ServerContext::AddTrailingMetadata(const grpc::string& key,
 }
 
 void ServerContext::TryCancel() const {
-  internal::CancelInterceptorBatchMethods cancel_methods;
+  ::grpc::internal::CancelInterceptorBatchMethods cancel_methods;
   if (rpc_info_) {
     for (size_t i = 0; i < rpc_info_->interceptors_.size(); i++) {
       rpc_info_->RunInterceptor(&cancel_methods, i);
@@ -398,4 +400,4 @@ void ServerContext::SetLoadReportingCosts(
   }
 }
 
-}  // namespace grpc
+}  // namespace grpc_impl
