+++ test/cpp/codegen/compiler_test_golden
@@ -43,6 +43,7 @@
 namespace grpc_impl {
 class CompletionQueue;
 class ServerCompletionQueue;
+class ServerContext;
 }  // namespace grpc_impl
 
 namespace grpc {
@@ -50,7 +51,6 @@ namespace experimental {
 template <typename RequestT, typename ResponseT>
 class MessageAllocator;
 }  // namespace experimental
-class ServerContext;
 }  // namespace grpc
 
 namespace grpc {
